+++ graphql/src/index.js
@@ -1,8 +1,8 @@
 // Local.
 const mongo = require("./mongo");
 const reddit = require("./reddit");
-const Comment = require("./models/comment");
-const Submission = require("./models/submission");
+const graphQlSchema = require("./schema/index");
+const graphQlResolvers = require("./resolvers/index");
 
 // Libraries
 const mongoose = require("mongoose");
@@ -10,7 +10,6 @@ const express = require("express");
 const app = express();
 const bodyParser = require("body-parser");
 const { graphqlHTTP } = require("express-graphql");
-const { buildSchema } = require("graphql");
 
 const db_uri =
   mongo["protocol"] +
@@ -26,76 +25,8 @@ app.use(bodyParser.json());
 app.use(
   "/graphql",
   graphqlHTTP({
-    schema: buildSchema(`
-      type Comment {
-        type: String!
-        author: String!
-        clicked: Boolean!
-        created_utc: Int!
-        distinguished: String
-        edited: Boolean!
-        reddit_id: String!
-        is_submitter: Boolean!
-        link_id: String!
-        parent_id: String!
-        permalink: String!
-        comment_link: String!
-        saved: Boolean!
-        score: Int!
-        stickied: Boolean!
-        subreddit: String!
-        subreddit_id: String!
-      }
-
-      type Submission {
-        type: String!
-        author: String!
-        clicked: Boolean!
-        created_utc: Int!
-        distinguished: String
-        edited: Boolean!
-        reddit_id: String!
-        is_original_content: Boolean!
-        is_self: Boolean!
-        link_flair_text: String!
-        locked: Boolean!
-        name: String!
-        num_comments: Int!
-        over_18: Boolean!
-        permalink: String!
-        submission_link: String!
-        saved: Boolean!
-        score: Int!
-        selftext: String!
-        spoiler: Boolean!
-        stickied: Boolean!
-        subreddit: String!
-        subreddit_id: String!
-        title: String!
-        upvote_ratio: Float!
-        url: String!
-      }
-
-      type RootQuery {
-        comments: [Comment]
-        submissions: [Submission]
-      }
-      schema {
-        query: RootQuery
-      }
-    `),
-    rootValue: {
-      comments: () => {
-        return Comment.find().then((result) => {
-          return result;
-        });
-      },
-      submissions: () => {
-        return Submission.find().then((result) => {
-          return result;
-        });
-      },
-    },
+    schema: graphQlSchema,
+    rootValue: graphQlResolvers,
     graphiql: true,
   })
 );
