+++ graphql/src/resolvers/index.js
@@ -1,0 +1,9 @@
+const commentsResolver = require("./comments");
+const submissionsResolver = require("./submissions");
+
+const rootResolver = {
+  ...commentsResolver,
+  ...submissionsResolver,
+};
+
+module.exports = rootResolver;
