+++ graphql/src/resolvers/submissions.js
@@ -1,0 +1,18 @@
+const Submission = require("../models/submission");
+
+module.exports = {
+  submissions: async () => {
+    try {
+      const submissions = await Submission.find();
+      return submissions;
+    } catch (err) {
+      throw err;
+    }
+  },
+};
+
+// submissions: () => {
+//   return Submission.find().then((result) => {
+//     return result;
+//   });
+// },
